requirements.txt
asgiref==3.3.0
boto3==1.16.18
botocore==1.19.18
Django==3.1.3
django-crispy-forms==1.9.2
django-filter==2.4.0
django-rest-framework==0.1.0
django-storages==1.10.1
djangorestframework==3.12.2
gunicorn==20.0.4
jmespath==0.10.0
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.4
redis==3.5.3
s3transfer==0.3.3
scipy==1.5.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
